requirements.txt
SpeechRecognition
pyaudio
pyttsx3
pyautogui
win10toast
requests
Pillow
infi.systray
pystray
